fix : models -> posts, users 이름 수정
